* Refactor/move startup related commands into controller/startup * Refactor/extract placeholder ipcMain handlers from PrepareViewCommand.ts into PrefsCommand.ts * Add PrepareControllerCommand.ts as async command that runs subcommands from api folder, each of which will handle ipcMain requests for a section of the api
